+++ pytest.ini
@@ -1,0 +1,22 @@
+[tool:pytest]
+# Register custom markers
+markers =
+    unit: Unit tests (fast, mocked dependencies)
+    integration: Integration tests (slow, real APIs)
+    slow: Slow tests that should be run separately
+    
+# Test discovery patterns
+testpaths = tests
+python_files = test_*.py
+python_classes = Test*
+python_functions = test_*
+
+# Filtering options
+addopts = 
+    -v
+    --tb=short
+    --strict-markers
+    
+# Default to exclude slow tests
+filterwarnings =
+    ignore::pytest.PytestUnknownMarkWarning
