+++ tests/unit/test_filter_df.py
@@ -4,7 +4,7 @@ import sys
 import os
 
 # Add project root to Python path
-sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
 
 from app.filter_df import apply_pre_filters, apply_post_filters, tone_options
